Move source_users & source_users_test code into sources.go & sources_test.go
